Update functions on block diagonal matrix and basis class for polar 2D

+++ tests/test_batched_covar2d.py
@@ -27,7 +27,8 @@ class BatchedRotCov2DTestCase(TestCase):
         filters = [RadialCTFFilter(pixel_size, voltage, defocus=d, Cs=2.0, alpha=0.1)
                    for d in np.linspace(defocus_min, defocus_max, defocus_ct)]
 
-        src = Simulation(L, n, filters=filters)
+        # Since FFBBasis2D doesn't yet implement dtype, we'll set this to double to match its built in types.
+        src = Simulation(L, n, filters=filters, dtype='double')
 
         basis = FFBBasis2D((L, L))
 
@@ -37,7 +38,7 @@ class BatchedRotCov2DTestCase(TestCase):
         ctf_fb = [f.fb_mat(basis) for f in unique_filters]
 
         im = src.images(0, src.n)
-        coeff = basis.evaluate_t(im.data)
+        coeff = basis.evaluate_t(im.data).astype(src.dtype)
 
         cov2d = RotCov2D(basis)
         bcov2d = BatchedRotCov2D(src, basis, batch_size=7)
@@ -62,7 +63,7 @@ class BatchedRotCov2DTestCase(TestCase):
         return close
 
     def test01(self):
-        # Test basic functionality again RotCov2D.
+        # Test basic functionality against RotCov2D.
         noise_var = 0.1848
 
         mean_cov2d = self.cov2d.get_mean(self.coeff, ctf_fb=self.ctf_fb,
